refatoracao do codigo

+++ src/main/java/com/grupomateus/triboraizes/desafio/service/ClienteService.java
@@ -1,17 +1,22 @@
 package com.grupomateus.triboraizes.desafio.service;
 
 import com.grupomateus.triboraizes.desafio.dto.ClienteDto;
+import com.grupomateus.triboraizes.desafio.exceptions.ClienteNaoEncontradoException;
 import com.grupomateus.triboraizes.desafio.mapper.ClienteMapper;
 import com.grupomateus.triboraizes.desafio.model.Cliente;
 import com.grupomateus.triboraizes.desafio.repository.ClienteRepository;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ClienteService {
-    private ClienteRepository clienteRepository;
 
+    private final ClienteRepository clienteRepository;
+
+    @Autowired
     public ClienteService(ClienteRepository clienteRepository) {
         this.clienteRepository = clienteRepository;
     }
@@ -23,18 +28,24 @@ public class ClienteService {
     }
 
     public ClienteDto buscarClientePorId(Long id) {
-        Cliente cliente = clienteRepository.findById(id).orElse(null);
-        if (cliente == null) {
-            return null;
+        Optional<Cliente> clienteOpt = clienteRepository.findById(id);
+        if (clienteOpt.isEmpty()) {
+            throw new ClienteNaoEncontradoException("Cliente não encontrado com o ID: " + id);
         }
-        return ClienteMapper.toClienteDto(cliente);
+        return ClienteMapper.toClienteDto(clienteOpt.get());
     }
 
     public void deletarClientePorId(Long id) {
+        if (!clienteRepository.existsById(id)) {
+            throw new ClienteNaoEncontradoException("Cliente não encontrado com o ID: " + id);
+        }
         clienteRepository.deleteById(id);
     }
 
     public ClienteDto atualizarCliente(ClienteDto clienteDto) {
+        if (!clienteRepository.existsById(clienteDto.id())) {
+            throw new ClienteNaoEncontradoException("Cliente não encontrado com o ID: " + clienteDto.id());
+        }
         Cliente cliente = ClienteMapper.toCliente(clienteDto);
         Cliente clienteSalvo = clienteRepository.save(cliente);
         return ClienteMapper.toClienteDto(clienteSalvo);
